Polish test push notification
